Revamp snap point implementation to fix multiple issues
Change the way snap points are calculated. Snap points are now correctly calculated from the sheet height instead of the window height. This makes it easier to reason about snap points.

Additionally, the opening animation of the sheet has been moved from `AnimatePresence` to a `useEffect` with directly calling `animate` to allow reading the sheet height that is needed for calculating the snap points in the new approach.